Output on vertices complete
Output on vertices now validated (against element results with consolidation cube Ehlers problem). NOTE: both element and vertex results are NOT WEIGHTED, but rather added up for each element/vertex and divided by number of GP contributions. 
Load on surface and seepage velocity in whole mesh when averaged over nodes are output as a vector that can be interpreted and plotted as such by Paraview.